+++ changelog.md
@@ -5,6 +5,8 @@ Rails-i18nterface Changelog
 
 ### v0.2.1 :
 
+* added a per_page input on view
+* moved more methods from controller to libs
 * added a cache (using marshall) for teh scan of translatable string extracted from application
 
 ### v0.2.0 :